HS2.0: put wifi config in to current user map if added from system
According to current flow, the wifi config can be added to current user
map if it create by wifi group or it is a shared config. However, if we
want to connect to a AP supported Hotspot 2.0, the config wouldn't be
shared and if it added by system group, it will fail to put into
current user map then always fail to choice. So add adjustment allow it
added from system like shell commands.

CRs-Fixed: 3026295
Change-Id: I1fd3b9319eec7310734a8cacec6dedcc3c570865