p2p: send tethering request with TETHER_PRIVILEGED permission
On T or later, a broadcast with FINE_LOCATION requirement cannot be be
received by a receiver with FINE_LOCATION when location is off.

As the tethering request is a dedicate broadcast, it is better to use
TETHER_PRIVILEGED permission to neglect location mode state for
internal service communication.

Bug: 229183910
Test: create a group when location is off
      atest FrameworksWifiTests
Change-Id: Ic091a52b03d81234e15046d7cd1e82d0b5722b3f

(cherry picked from commit 9edb7973e607943258b91188afa032e4a975e1ec)